KO-193 Add ad hoc job build and deploy job (#142)
Add ad hoc job build and deploy job

This job is useful for running e2e tests against a feature branch. It will
build the branch, deploy the artifact to artifactory, and run the e2e tests
just as if it were a stable branch. The primary reason for having a separate
job for ad hoc builds is to prevent the build run logs in jenkins from getting
polluted with billy bob's feature of madness logs.