requeue after the last node has its node-state label set to Started
This commit fixes an issue in which we basically see out of order status
updates. Specifically, the .Status.SuperUserUpserted property is getting set
before the last node is added to .Status.NodeStatuses. The actual order of
operations is correct though. The super user is not created until all C* nodes
are started. It is just that the status updates basically occur out of order.