Mount root as read-only and temp dir as memory emptyvol #86 (#89)
* Mount root as read-only and temp dir as memory emptyvol #86

* Update service port on chart files so operator can run as nobody (#86)

* Use the explicitly provisioned memory tmpdir for temporary cert location

The directory that was used prior to this commit is created automatically as
a byproduct of mounting the secret under the tmp volume which we explicitly
mount as a temporary dir. While the one we create is o+rwx, the one that is
implicitly created when the secret is mounted into the subdirectory is writable
only by root. This is only a problem now that we are running the operator as an
unprivileged user.