bump mgmt api image version
Switch to used mgmt api image v0.1.4 to get the fix for
https://github.com/datastax/management-api-for-apache-cassandra/issues/18 which
was blocking me. Without the fix, I could not deploy a CassandraDatacenter and
then enable Reaper.

Due to the changes in
https://github.com/datastax/management-api-for-apache-cassandra/commit/1a225f99f91863673b074b3c1a9c1d2d9cd79ecd#diff-3aaf645b2133cebc86740ab63150b507
I also had to add the IGNORE_DEFAULTS=true env var to the cassandra container
in order for seeds to get configured properly.