initial support for using C* authentication with reaper
Before Reaper can successfully deploy, Cassandra roles have to be created and
the reaper keyspace needs to be initialized. When the statefulset is created,
the reaper container is not deployed initially anymore. When the schema init
job completes, the new ReaperStatus.SchemaInitialized field is set to true. On
the subsequent reconciliation loop the statefulset(s) are updated to deploy the
reaper container.