update replication_factor of system_auth keyspace
This commit adds a function for altering a keyspace which I needed to configure
the RF of system_auth. When authentication is enabled, system_auth should never
have RF = 1. Because the amount of data is very small, increase the RF even
when authentication is disabled is not a big deal.

In my testing for reaper sidecar, I found that the RF of system_auth has to be
changed when authentication is enabled; otherwise, there will be lots of
flapping of containers as only one C* node will have the credentials and it
might not the node that is up.