deploy a second job to apply reaper schema changes
Reaper coordinates schema changes with distributed locks implemented in a C*
table. This is done in order to avoid concurrent schema updates which is
problematic in C*. The locks table uses IP addresses. When (not if) the reaper
sidecar container liveness and/or readiness probe fails and the container
restarts, it will come back up with a differnet IP address. Reaper will then
have to wait to reacquire the lock which can lead to additional liveness and/or
readiness probe failures. By applying schema updates with a single instance via
a job, these problems go away.

Lastly, based on this as well as on previous experience, I firmly believe that
C* schema changes in a k8s deployment should always be done in a k8s job to
avoid these types of problems.